requirements.txt
numpy==1.26.3
opencv-python==4.9.0.80
pynrrd==1.0.0
tifffile==2024.2.12
tqdm==4.66.2
